OID4VCI: Extract separate authorization service
